#6117 feat(refactor): refactors GridCell component
