+++ src/components/Grid/GridCell/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GridCell';
